pyproject.toml
[project]
name = "agentic-ai-architectures"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "langchain-groq>=1.1.1",
    "langchain-tavily>=0.2.16",
    "langgraph>=1.0.5",
    "notebook>=7.5.1",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]
